LayoutTests/platform/gtk-wk2/accessibility/file-upload-button-stringvalue-expected.txt

  
This tests the value of stringValue for a single file- and multiple files-<input type="file"> control.


Default value (no files selected):
accessibilityController.accessibleElementById("singleFileUpload").stringValue: AXValue: Choose File
accessibilityController.accessibleElementById("multipleFileUpload").stringValue: AXValue: Choose Files

Drag and drop a single file:
accessibilityController.accessibleElementById("singleFileUpload").stringValue: AXValue: Choose File
accessibilityController.accessibleElementById("multipleFileUpload").stringValue: AXValue: Choose Files

Drag and drop two files:
accessibilityController.accessibleElementById("singleFileUpload").stringValue: AXValue: Choose File
accessibilityController.accessibleElementById("multipleFileUpload").stringValue: AXValue: Choose Files
